QSP-8 Retrieved Collection Index May Represent Chain Index

+++ src/solc_0.8/assetERC1155/AssetBaseERC1155.sol
@@ -214,6 +214,8 @@ abstract contract AssetBaseERC1155 is WithSuperOperators, IERC1155 {
             id == 0x572b6c05; // ERC2771
     }
 
+    /// Collection methods for ERC721s extracted from an ERC1155 -----------------------------------------------------
+
     /// @notice Gives the collection a specific token belongs to.
     /// @param id the token to get the collection of.
     /// @return the collection the NFT is part of.
@@ -237,8 +239,10 @@ abstract contract AssetBaseERC1155 is WithSuperOperators, IERC1155 {
     /// @return the index/order at which the token `id` was minted in a collection.
     function collectionIndexOf(uint256 id) external view returns (uint256) {
         collectionOf(id); // this check if id and collection indeed was ever minted
-        return uint32((id & ERC1155ERC721Helper.NFT_INDEX) >> ERC1155ERC721Helper.NFT_INDEX_OFFSET);
+        return uint24((id & ERC1155ERC721Helper.NFT_INDEX) >> ERC1155ERC721Helper.NFT_INDEX_OFFSET);
     }
+
+    /// end collection methods ---------------------------------------------------------------------------------------
 
     function wasEverMinted(uint256 id) public view returns (bool) {
         return _metadataHash[id & ERC1155ERC721Helper.URI_ID] != 0;
@@ -277,14 +281,14 @@ abstract contract AssetBaseERC1155 is WithSuperOperators, IERC1155 {
         require(isBouncer(_msgSender()), "!BOUNCER");
         require(to != address(0), "TO==0");
         require(id & ERC1155ERC721Helper.IS_NFT == 0, "UNIQUE_ERC1155");
-        uint32 tokenCollectionIndex = _nextCollectionIndex[id] + 1;
+        uint24 tokenCollectionIndex = uint24(_nextCollectionIndex[id]) + 1;
         _nextCollectionIndex[id] = tokenCollectionIndex;
         string memory metaData = tokenURI(id);
         uint256 newId =
             id +
-                ERC1155ERC721Helper.IS_NFT_OFFSET_MULTIPLIER +
+                ERC1155ERC721Helper.IS_NFT_OFFSET_MULTIPLIER + // newId is always an NFT; IS_NFT is 1
                 (tokenCollectionIndex) *
-                2**ERC1155ERC721Helper.NFT_INDEX_OFFSET;
+                2**ERC1155ERC721Helper.NFT_INDEX_OFFSET; // uint24 nft index
         _burnFT(sender, id, 1);
         _assetERC721.mint(to, newId, bytes(abi.encode(metaData)));
         emit Extraction(id, newId);
